🚹 Resource based access added to project controllers

+++ src/projects/projects.controller.ts
@@ -1,8 +1,21 @@
-import { Body, Controller, Delete, Get, Param, Patch, Post } from '@nestjs/common';
+import {
+  Body,
+  ConflictException,
+  Controller,
+  Delete,
+  ForbiddenException,
+  Get,
+  Param,
+  Patch,
+  Post,
+} from '@nestjs/common';
+import { CurrentUser } from 'src/core/decorators/current-user.decorator';
+import { ProjectRole } from 'src/core/enums/project-role.enum';
 import { ProjectMemberDto } from 'src/project-members/dtos/project-member.dto';
 import { CreateProjectTaskDto } from 'src/project-tasks/dto/create-project-task.dto';
 import { ProjectTaskDto } from 'src/project-tasks/dto/project-task.dto';
 import { UpdateProjectTaskDto } from 'src/project-tasks/dto/update-project-task.dto';
+import { User } from 'src/users/schemas/user.schema';
 import { CreateProjectMemberDto } from '../project-members/dtos/create-project-member.dto';
 import { ProjectDto } from './dto/project.dto';
 import { ProjectsService } from './projects.service';
@@ -12,55 +25,112 @@ export class ProjectsController {
   constructor(private readonly projectsService: ProjectsService) {}
 
   @Get(':id')
-  findOne(@Param('id') id: string): Promise<ProjectDto> {
-    return this.projectsService.findOne(id);
+  async findOne(@Param('id') id: string, @CurrentUser() user): Promise<ProjectDto> {
+    const project = await this.projectsService.findOne(id);
+    if (!project.members.find(m => m.user === user.sub)) {
+      throw new ForbiddenException('You are not allowed to access this project');
+    }
+    return project;
   }
 
   @Get('/:id/tasks/:taskId')
-  findTask(@Param('id') id: string, @Param('taskId') taskId: string): Promise<ProjectTaskDto> {
+  async findTask(
+    @Param('id') id: string,
+    @Param('taskId') taskId: string,
+    @CurrentUser() user
+  ): Promise<ProjectTaskDto> {
+    const project = await this.projectsService.findOne(id);
+    if (!project.members.find(m => m.user === user.sub)) {
+      throw new ForbiddenException('You are not allowed to access this task on this project');
+    }
     return this.projectsService.findTask(id, taskId);
   }
 
   @Post('/:id/tasks')
-  createTask(@Param('id') id: string, @Body() createProjectTaskDto: CreateProjectTaskDto): Promise<ProjectTaskDto> {
+  async createTask(
+    @Param('id') id: string,
+    @Body() createProjectTaskDto: CreateProjectTaskDto,
+    @CurrentUser() user
+  ): Promise<ProjectTaskDto> {
+    const project = await this.projectsService.findOne(id);
+    if (!project.members.find(m => m.user === user.sub && m.role === (ProjectRole.OWNER || ProjectRole.CLIENT))) {
+      throw new ForbiddenException('You are not allowed to create a task on this project');
+    }
     return this.projectsService.createTask(id, createProjectTaskDto);
   }
 
   @Patch('/:id/tasks/:taskId')
-  updateTask(
+  async updateTask(
     @Param('id') id: string,
     @Param('taskId') taskId: string,
-    @Body() updateProjectTaskDto: UpdateProjectTaskDto
+    @Body() updateProjectTaskDto: UpdateProjectTaskDto,
+    @CurrentUser() user
   ): Promise<ProjectTaskDto> {
+    const project = await this.projectsService.findOne(id);
+    if (!project.members.find(m => m.user === user.sub && m.role === (ProjectRole.OWNER || ProjectRole.CLIENT))) {
+      throw new ForbiddenException('You are not allowed to update tasks on this project');
+    }
     return this.projectsService.updateTask(id, taskId, updateProjectTaskDto);
   }
 
   @Delete('/:id/tasks/:taskId')
-  deleteTask(@Param('id') id: string, @Param('taskId') taskId: string): Promise<string> {
+  async deleteTask(@Param('id') id: string, @Param('taskId') taskId: string, @CurrentUser() user): Promise<string> {
+    const project = await this.projectsService.findOne(id);
+    if (!project.members.find(m => m.user === user.sub && m.role === ProjectRole.OWNER)) {
+      throw new ForbiddenException('You are not allowed to delete tasks on this project');
+    }
     return this.projectsService.deleteTask(id, taskId);
   }
 
   @Get(':id/tasks')
-  listTasks(@Param('id') id: string): Promise<ProjectTaskDto[]> {
+  async listTasks(@Param('id') id: string, @CurrentUser() user): Promise<ProjectTaskDto[]> {
+    const project = await this.projectsService.findOne(id);
+    if (!project.members.find(m => m.user === user.sub)) {
+      throw new ForbiddenException('You are not allowed to see the tasks on this project');
+    }
     return this.projectsService.listTasks(id);
   }
 
   @Post(':id/members/:userEmail')
-  inviteUser(
+  async inviteUser(
     @Param('id') id: string,
     @Param('userEmail') userEmail: string,
-    @Body() createProjectUserDto: CreateProjectMemberDto
+    @Body() createProjectUserDto: CreateProjectMemberDto,
+    @CurrentUser() user
   ) {
+    if (user.email === userEmail) {
+      throw new ConflictException('You cannot invite yourself to a project');
+    }
+
+    const project = await this.projectsService.findOne(id);
+
+    if (!project.members.find(m => m.user === user.sub && m.role === ProjectRole.OWNER)) {
+      throw new ForbiddenException('You are not allowed to invite users to this project');
+    }
     return this.projectsService.inviteUser(id, userEmail, createProjectUserDto);
   }
 
   @Delete(':id/members/:userId')
-  uninviteUser(@Param('id') id: string, @Param('userId') userId: string) {
+  async uninviteUser(@Param('id') id: string, @Param('userId') userId: string, @CurrentUser() user) {
+    const project = await this.projectsService.findOne(id);
+    if (!project.members.find(m => m.user === user.sub && m.role === ProjectRole.OWNER)) {
+      throw new ForbiddenException('You are not allowed to uninvite users from this project');
+    }
+    if (!project.members.find(m => m.user === userId)) {
+      throw new ConflictException('User is not part of this project');
+    }
+    if (project.members.find(m => m.user === userId && m.role === ProjectRole.OWNER)) {
+      throw new ForbiddenException('You are not allowed to uninvite an owner from a project');
+    }
     return this.projectsService.uninviteUser(id, userId);
   }
 
   @Get(':id/members/listInvitedUsers')
-  listInvitedUsers(@Param('id') id: string): Promise<ProjectMemberDto[]> {
+  async listInvitedUsers(@Param('id') id: string, @CurrentUser() user): Promise<ProjectMemberDto[]> {
+    const project = await this.projectsService.findOne(id);
+    if (!project.members.find(m => m.user === user.sub)) {
+      throw new ForbiddenException('You are not allowed to see the users for this project');
+    }
     return this.projectsService.listProjectMembers(id);
   }
 }
